+++ src/ProjNative/ProjNative.cpp
@@ -17,8 +17,30 @@ DllExport(PJ*) pCreateTransformation(PJ_CONTEXT* context, const char* src, const
 DllExport(void) pDestroyTransformation(PJ* transformation) {
     proj_destroy(transformation);
 }
+DllExport(PJ*) pCreateCoordinateSystem(PJ_CONTEXT* context, const char* definition) {
+    auto h = proj_create(context, definition);
+    if(proj_is_crs(h)) {
+        return h;
+    }
+    else {
+        proj_destroy(h);
+        char* def = new char[strlen(definition) + 50];
+        sprintf(def, "%s +type=crs", definition);
+        h = proj_create(context, def);
+        delete[] def;
+        return h;
+    }
+}
 
-DllExport(bool) pTransform(PJ* transformation, int count, V3d* points, V3d* outPoints) {
+DllExport(void) pDestroyCoordinateSystem(PJ* coordinateSystem) {
+    proj_destroy(coordinateSystem);
+}
+
+DllExport(PJ*) pCreateCoordinateTransformation(PJ_CONTEXT* context, PJ* src, PJ* dst, const char * const options[]) {
+    return proj_create_crs_to_crs_from_pj(context, src, dst, NULL, options);
+}
+
+DllExport(bool) pTransformArray(PJ* transformation, int count, V3d* points, V3d* outPoints) {
     
     PJ_COORD* result = new PJ_COORD[count];
     for(int i = 0; i < count; i++) {
@@ -33,6 +55,26 @@ DllExport(bool) pTransform(PJ* transformation, int count, V3d* points, V3d* outP
         outPoints[i].Y = result[i].xyz.y;
         outPoints[i].Z = result[i].xyz.z;
     }
+    delete[] result;
 
+    return err == 0;
+}
+
+DllExport(bool) pTransformArray2d(PJ* transformation, int count, V2d* points, V2d* outPoints) {
+    
+    PJ_COORD* result = new PJ_COORD[count];
+    for(int i = 0; i < count; i++) {
+        result[i].xyz.x = points[i].X;
+        result[i].xyz.y = points[i].Y;
+        result[i].xyz.z = 0.0;
+    }
+
+    auto err = proj_trans_array(transformation, PJ_FWD, count, result);
+    for(int i = 0; i < count; i++) {
+        outPoints[i].X = result[i].xyz.x;
+        outPoints[i].Y = result[i].xyz.y;
+    }
+    delete[] result;
+    
     return err == 0;
 }