+++ RELEASE_NOTES.md
@@ -1,2 +1,6 @@
+### 0.0.2
+* explicit CoordinateSystems
+* static API via Proj.Transform/etc.
+
 ### 0.0.1
 * initial version